LPD-20678 Move package, remove portal.settings
