+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/CategorizationSiteSettingsConfigurationScreenWrapper.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
 import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
@@ -22,13 +23,13 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class LegacyEditSiteSettingsConfigurationScreenWrapper
+public class CategorizationSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new LegacyEditSiteSettingsConfigurationScreenContributor());
+			new CategorizationSiteSettingsConfigurationScreenContributor());
 	}
 
 	@Reference
@@ -41,32 +42,46 @@ public class LegacyEditSiteSettingsConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class LegacyEditSiteSettingsConfigurationScreenContributor
+	private class CategorizationSiteSettingsConfigurationScreenContributor
 		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "other";
+			return "assets";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/site.jsp";
+			return "/site_settings/categorization.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-other";
+			return "site-configuration-categorization";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "other");
+			return _language.get(locale, "categorization");
+		}
+
+		@Override
+		public String getSaveMVCActionCommandName() {
+			return "/site_admin/edit_categorization";
 		}
 
 		@Override
 		public ServletContext getServletContext() {
 			return _servletContext;
+		}
+
+		@Override
+		public boolean isVisible(Group group) {
+			if (group.isCompany()) {
+				return false;
+			}
+
+			return true;
 		}
 
 	}
