+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/DefaultUserAssociationsSiteSettingsConfigurationScreenWrapper.java
@@ -3,18 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
+import com.liferay.item.selector.ItemSelector;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.LayoutSetPrototype;
-import com.liferay.portal.kernel.service.LayoutSetPrototypeLocalService;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.site.admin.web.internal.display.context.DefaultUserAssociationsDisplayContext;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
@@ -22,6 +18,7 @@ import java.util.Locale;
 
 import javax.servlet.ServletContext;
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -30,20 +27,20 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class PublicPrivatePagesSiteSettingsConfigurationScreenWrapper
+public class DefaultUserAssociationsSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new PublicPrivatePagesSiteSettingsConfigurationScreenContributor());
+			new DefaultUserAssociationsSiteSettingsConfigurationScreenContributor());
 	}
 
 	@Reference
-	private Language _language;
+	private ItemSelector _itemSelector;
 
 	@Reference
-	private LayoutSetPrototypeLocalService _layoutSetPrototypeLocalService;
+	private Language _language;
 
 	@Reference(target = "(osgi.web.symbolicname=com.liferay.site.admin.web)")
 	private ServletContext _servletContext;
@@ -52,32 +49,33 @@ public class PublicPrivatePagesSiteSettingsConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class PublicPrivatePagesSiteSettingsConfigurationScreenContributor
-		implements SiteSettingsConfigurationScreenContributor {
+	private class
+		DefaultUserAssociationsSiteSettingsConfigurationScreenContributor
+			implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "pages";
+			return "users";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/public_private_pages.jsp";
+			return "/site_settings/default_user_associations.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-public-private-pages";
+			return "site-configuration-default-user-associations";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "pages");
+			return _language.get(locale, "default-user-associations");
 		}
 
 		@Override
 		public String getSaveMVCActionCommandName() {
-			return "/site_admin/edit_pages";
+			return "/site_admin/edit_default_user_associations";
 		}
 
 		@Override
@@ -87,35 +85,25 @@ public class PublicPrivatePagesSiteSettingsConfigurationScreenWrapper
 
 		@Override
 		public boolean isVisible(Group group) {
-			if ((group != null) && group.isCompany()) {
+			if (group.isCompany()) {
 				return false;
 			}
 
-			if ((group != null) && group.isPrivateLayoutsEnabled()) {
-				return true;
-			}
+			return true;
+		}
 
-			ServiceContext serviceContext =
-				ServiceContextThreadLocal.getServiceContext();
+		@Override
+		public void setAttributes(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse) {
 
-			ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
+			SiteSettingsConfigurationScreenContributor.super.setAttributes(
+				httpServletRequest, httpServletResponse);
 
-			HttpServletRequest httpServletRequest = themeDisplay.getRequest();
-
-			long layoutSetPrototypeId = ParamUtil.getLong(
-				httpServletRequest, "layoutSetPrototypeId");
-
-			if (layoutSetPrototypeId > 0) {
-				LayoutSetPrototype layoutSetPrototype =
-					_layoutSetPrototypeLocalService.fetchLayoutSetPrototype(
-						layoutSetPrototypeId);
-
-				if (layoutSetPrototype != null) {
-					return true;
-				}
-			}
-
-			return false;
+			httpServletRequest.setAttribute(
+				DefaultUserAssociationsDisplayContext.class.getName(),
+				new DefaultUserAssociationsDisplayContext(
+					httpServletRequest, _itemSelector));
 		}
 
 	}
