+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/LanguagesSiteSettingsConfigurationScreenWrapper.java
@@ -3,56 +3,37 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
-import com.liferay.item.selector.ItemSelector;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.site.admin.web.internal.display.context.MenuAccessConfigurationDisplayContext;
-import com.liferay.site.configuration.manager.MenuAccessConfigurationManager;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
 import java.util.Locale;
 
 import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Mikel Lorza
+ * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class MenuAccessConfigurationScreenWrapper
+public class LanguagesSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new MenuAccessConfigurationScreenContributor());
+			new LanguagesSiteSettingsConfigurationScreenContributor());
 	}
 
 	@Reference
-	private ItemSelector _itemSelector;
-
-	@Reference
 	private Language _language;
-
-	@Reference
-	private MenuAccessConfigurationManager _menuAccessConfigurationManager;
-
-	@Reference
-	private Portal _portal;
-
-	@Reference
-	private RoleLocalService _roleLocalService;
 
 	@Reference(target = "(osgi.web.symbolicname=com.liferay.site.admin.web)")
 	private ServletContext _servletContext;
@@ -61,32 +42,32 @@ public class MenuAccessConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class MenuAccessConfigurationScreenContributor
+	private class LanguagesSiteSettingsConfigurationScreenContributor
 		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "site-configuration";
+			return "localization";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/menu_access.jsp";
+			return "/site_settings/languages.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-menu-access";
+			return "site-configuration-languages";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "menu-access");
+			return _language.get(locale, "languages");
 		}
 
 		@Override
 		public String getSaveMVCActionCommandName() {
-			return "/site_settings/edit_menu_access_configuration";
+			return "/site_admin/edit_languages";
 		}
 
 		@Override
@@ -101,19 +82,6 @@ public class MenuAccessConfigurationScreenWrapper
 			}
 
 			return true;
-		}
-
-		@Override
-		public void setAttributes(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse) {
-
-			httpServletRequest.setAttribute(
-				MenuAccessConfigurationDisplayContext.class.getName(),
-				new MenuAccessConfigurationDisplayContext(
-					httpServletRequest, _itemSelector,
-					_menuAccessConfigurationManager, _portal,
-					_roleLocalService));
 		}
 
 	}
