+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/MapsSiteSettingsConfigurationScreenWrapper.java
@@ -3,18 +3,25 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
+import com.liferay.map.constants.MapProviderWebKeys;
+import com.liferay.map.util.MapProviderHelperUtil;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
 import java.util.Locale;
 
 import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -23,14 +30,17 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class CategorizationSiteSettingsConfigurationScreenWrapper
+public class MapsSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new CategorizationSiteSettingsConfigurationScreenContributor());
+			new MapsSiteSettingsConfigurationScreenContributor());
 	}
+
+	@Reference
+	private GroupLocalService _groupLocalService;
 
 	@Reference
 	private Language _language;
@@ -42,32 +52,27 @@ public class CategorizationSiteSettingsConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class CategorizationSiteSettingsConfigurationScreenContributor
+	private class MapsSiteSettingsConfigurationScreenContributor
 		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "assets";
+			return "maps";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/categorization.jsp";
+			return "/site_settings/maps.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-categorization";
+			return "site-settings-maps";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "categorization");
-		}
-
-		@Override
-		public String getSaveMVCActionCommandName() {
-			return "/site_admin/edit_categorization";
+			return _language.get(locale, "maps");
 		}
 
 		@Override
@@ -76,12 +81,33 @@ public class CategorizationSiteSettingsConfigurationScreenWrapper
 		}
 
 		@Override
-		public boolean isVisible(Group group) {
-			if (group.isCompany()) {
-				return false;
+		public void setAttributes(
+			HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse) {
+
+			SiteSettingsConfigurationScreenContributor.super.setAttributes(
+				httpServletRequest, httpServletResponse);
+
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)httpServletRequest.getAttribute(
+					WebKeys.THEME_DISPLAY);
+
+			Group siteGroup = themeDisplay.getSiteGroup();
+
+			Group liveGroup = null;
+
+			if (siteGroup.isStagingGroup()) {
+				liveGroup = siteGroup.getLiveGroup();
+			}
+			else {
+				liveGroup = siteGroup;
 			}
 
-			return true;
+			httpServletRequest.setAttribute(
+				MapProviderWebKeys.MAP_PROVIDER_KEY,
+				MapProviderHelperUtil.getMapProviderKey(
+					_groupLocalService, themeDisplay.getCompanyId(),
+					liveGroup.getGroupId()));
 		}
 
 	}
