+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/PublicPrivatePagesSiteSettingsConfigurationScreenWrapper.java
@@ -3,21 +3,25 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.ratings.kernel.definition.PortletRatingsDefinitionUtil;
-import com.liferay.ratings.kernel.definition.PortletRatingsDefinitionValues;
+import com.liferay.portal.kernel.model.LayoutSetPrototype;
+import com.liferay.portal.kernel.service.LayoutSetPrototypeLocalService;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
 import java.util.Locale;
-import java.util.Map;
 
 import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -26,17 +30,20 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class RatingsSiteSettingsConfigurationScreenWrapper
+public class PublicPrivatePagesSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new RatingsSiteSettingsConfigurationScreenContributor());
+			new PublicPrivatePagesSiteSettingsConfigurationScreenContributor());
 	}
 
 	@Reference
 	private Language _language;
+
+	@Reference
+	private LayoutSetPrototypeLocalService _layoutSetPrototypeLocalService;
 
 	@Reference(target = "(osgi.web.symbolicname=com.liferay.site.admin.web)")
 	private ServletContext _servletContext;
@@ -45,32 +52,32 @@ public class RatingsSiteSettingsConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class RatingsSiteSettingsConfigurationScreenContributor
+	private class PublicPrivatePagesSiteSettingsConfigurationScreenContributor
 		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "community-tools";
+			return "pages";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/ratings.jsp";
+			return "/site_settings/public_private_pages.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-ratings";
+			return "site-configuration-public-private-pages";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "ratings");
+			return _language.get(locale, "pages");
 		}
 
 		@Override
 		public String getSaveMVCActionCommandName() {
-			return "/site_admin/edit_ratings";
+			return "/site_admin/edit_pages";
 		}
 
 		@Override
@@ -80,16 +87,35 @@ public class RatingsSiteSettingsConfigurationScreenWrapper
 
 		@Override
 		public boolean isVisible(Group group) {
-			if (group.isCompany()) {
+			if ((group != null) && group.isCompany()) {
 				return false;
 			}
 
-			Map<String, PortletRatingsDefinitionValues>
-				portletRatingsDefinitionValuesMap =
-					PortletRatingsDefinitionUtil.
-						getPortletRatingsDefinitionValuesMap();
+			if ((group != null) && group.isPrivateLayoutsEnabled()) {
+				return true;
+			}
 
-			return !portletRatingsDefinitionValuesMap.isEmpty();
+			ServiceContext serviceContext =
+				ServiceContextThreadLocal.getServiceContext();
+
+			ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
+
+			HttpServletRequest httpServletRequest = themeDisplay.getRequest();
+
+			long layoutSetPrototypeId = ParamUtil.getLong(
+				httpServletRequest, "layoutSetPrototypeId");
+
+			if (layoutSetPrototypeId > 0) {
+				LayoutSetPrototype layoutSetPrototype =
+					_layoutSetPrototypeLocalService.fetchLayoutSetPrototype(
+						layoutSetPrototypeId);
+
+				if (layoutSetPrototype != null) {
+					return true;
+				}
+			}
+
+			return false;
 		}
 
 	}
