+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/RatingsSiteSettingsConfigurationScreenWrapper.java
@@ -3,22 +3,21 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
-import com.liferay.item.selector.ItemSelector;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.site.admin.web.internal.display.context.DefaultUserAssociationsDisplayContext;
+import com.liferay.ratings.kernel.definition.PortletRatingsDefinitionUtil;
+import com.liferay.ratings.kernel.definition.PortletRatingsDefinitionValues;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
 import java.util.Locale;
+import java.util.Map;
 
 import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -27,17 +26,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class DefaultUserAssociationsSiteSettingsConfigurationScreenWrapper
+public class RatingsSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new DefaultUserAssociationsSiteSettingsConfigurationScreenContributor());
+			new RatingsSiteSettingsConfigurationScreenContributor());
 	}
-
-	@Reference
-	private ItemSelector _itemSelector;
 
 	@Reference
 	private Language _language;
@@ -49,33 +45,32 @@ public class DefaultUserAssociationsSiteSettingsConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class
-		DefaultUserAssociationsSiteSettingsConfigurationScreenContributor
-			implements SiteSettingsConfigurationScreenContributor {
+	private class RatingsSiteSettingsConfigurationScreenContributor
+		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "users";
+			return "community-tools";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/default_user_associations.jsp";
+			return "/site_settings/ratings.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-default-user-associations";
+			return "site-configuration-ratings";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "default-user-associations");
+			return _language.get(locale, "ratings");
 		}
 
 		@Override
 		public String getSaveMVCActionCommandName() {
-			return "/site_admin/edit_default_user_associations";
+			return "/site_admin/edit_ratings";
 		}
 
 		@Override
@@ -89,21 +84,12 @@ public class DefaultUserAssociationsSiteSettingsConfigurationScreenWrapper
 				return false;
 			}
 
-			return true;
-		}
+			Map<String, PortletRatingsDefinitionValues>
+				portletRatingsDefinitionValuesMap =
+					PortletRatingsDefinitionUtil.
+						getPortletRatingsDefinitionValuesMap();
 
-		@Override
-		public void setAttributes(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse) {
-
-			SiteSettingsConfigurationScreenContributor.super.setAttributes(
-				httpServletRequest, httpServletResponse);
-
-			httpServletRequest.setAttribute(
-				DefaultUserAssociationsDisplayContext.class.getName(),
-				new DefaultUserAssociationsDisplayContext(
-					httpServletRequest, _itemSelector));
+			return !portletRatingsDefinitionValuesMap.isEmpty();
 		}
 
 	}
