+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/configuration/admin/display/RecycleBinSiteSettingsConfigurationScreenWrapper.java
@@ -3,12 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.site.admin.web.internal.portal.settings.configuration.admin.display;
+package com.liferay.site.admin.web.internal.configuration.admin.display;
 
 import com.liferay.configuration.admin.display.ConfigurationScreen;
 import com.liferay.configuration.admin.display.ConfigurationScreenWrapper;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.util.PrefsPropsUtil;
+import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenContributor;
 import com.liferay.site.settings.configuration.admin.display.SiteSettingsConfigurationScreenFactory;
 
@@ -23,13 +25,13 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(service = ConfigurationScreen.class)
-public class PublicPrivateSiteURLSiteSettingsConfigurationScreenWrapper
+public class RecycleBinSiteSettingsConfigurationScreenWrapper
 	extends ConfigurationScreenWrapper {
 
 	@Override
 	protected ConfigurationScreen getConfigurationScreen() {
 		return _siteSettingsConfigurationScreenFactory.create(
-			new PublicPrivateSiteURLSiteSettingsConfigurationScreenContributor());
+			new RecycleBinSiteSettingsConfigurationScreenContributor());
 	}
 
 	@Reference
@@ -42,32 +44,32 @@ public class PublicPrivateSiteURLSiteSettingsConfigurationScreenWrapper
 	private SiteSettingsConfigurationScreenFactory
 		_siteSettingsConfigurationScreenFactory;
 
-	private class PublicPrivateSiteURLSiteSettingsConfigurationScreenContributor
+	private class RecycleBinSiteSettingsConfigurationScreenContributor
 		implements SiteSettingsConfigurationScreenContributor {
 
 		@Override
 		public String getCategoryKey() {
-			return "site-configuration";
+			return "recycle-bin";
 		}
 
 		@Override
 		public String getJspPath() {
-			return "/site_settings/public_private_site_url.jsp";
+			return "/site_settings/recycle_bin.jsp";
 		}
 
 		@Override
 		public String getKey() {
-			return "site-configuration-public-private-site-url";
+			return "site-configuration-recycle-bin";
 		}
 
 		@Override
 		public String getName(Locale locale) {
-			return _language.get(locale, "site-url");
+			return _language.get(locale, "recycle-bin");
 		}
 
 		@Override
 		public String getSaveMVCActionCommandName() {
-			return "/site_admin/edit_site_url";
+			return "/site_admin/edit_recycle_bin";
 		}
 
 		@Override
@@ -77,7 +79,10 @@ public class PublicPrivateSiteURLSiteSettingsConfigurationScreenWrapper
 
 		@Override
 		public boolean isVisible(Group group) {
-			if (group.isCompany() || !group.isPrivateLayoutsEnabled()) {
+			boolean trashEnabled = PrefsPropsUtil.getBoolean(
+				group.getCompanyId(), PropsKeys.TRASH_ENABLED);
+
+			if (!trashEnabled) {
 				return false;
 			}
